refactor: isolate default runtime composition
